separate markdown parser
